backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["MoAlkhateeb <mo.hany.alkhateeb@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = "^0.34.0"
pydantic = "^2.10.6"
sqlalchemy = "^2.0.39"
pyjwt = "^2.10.1"
bcrypt = "^4.3.0"
stripe = "^12.0.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
